arc: fix arckey value

+++ files/initrd/opt/arc/include/functions.sh
@@ -645,11 +645,12 @@ function systemCheck () {
     CPUFREQ="false"
   fi
   # Check for ARCKEY
+  ARCKEY="$(readConfigKey "arc.key" "${USER_CONFIG_FILE}")"
   if openssl enc -in "${S_FILE_ENC}" -out "${S_FILE_ARC}" -d -aes-256-cbc -k "${ARCKEY}" 2>/dev/null; then
     cp -f "${S_FILE_ARC}" "${S_FILE}"
     writeConfigKey "arc.key" "${ARCKEY}" "${USER_CONFIG_FILE}"
   else
-    cp -f "${S_FILE}" "${S_FILE}.bak"
+    [ -f "${S_FILE}.bak" ] && cp -f "${S_FILE}" "${S_FILE}.bak"
     writeConfigKey "arc.key" "" "${USER_CONFIG_FILE}"
     writeConfigKey "arc.patch" "false" "${USER_CONFIG_FILE}"
   fi
